+++ internal/jlinkdll/env.go
@@ -1,0 +1,7 @@
+package jlinkdll
+
+import "os"
+
+func environ() []string {
+	return os.Environ()
+}